requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
numpy
pypdf
python-docx
pickle-mixin